refactor: remove dependency cycle
